added operation schedule button and moved variable newVM to a newVM.js file
