+++ app/views/infrastructures/_show.html.erb
@@ -124,3 +124,6 @@
 <script id="edit-attr-tabpane-template"       type="text/template"> <%= render 'vue/edit_attr_tabpane' %> </script>
 <script id="serverspec-tabpane-template"      type="text/template"> <%= render 'vue/serverspec_tabpane' %> </script>
 <script id="serverspec-results-tabpane-template"      type="text/template"> <%= render 'vue/serverspec_results_tabpane' %> </script>
+
+<!--Operation Schedule-->
+<script id="operation-sched-template" type="text/template"> <%=	render 'vue/operation_sched' %> </script>
